Fixes
 - Fixed the way retry/wait works: now, there will be a random wait at
the begining of each attempt to ensure that there won't be collisions on
the first attempt when multiple workers start at the same time, so
reducing subsequent retries. Wait time now depends on number of retries
so that the window will grow, reducing chance of collision. Also
increased number of retries.
 - Fixed error where test_mode was handled as boolean instead of string
 - If file move fails in test mode, exit with code 0 so workflow doesn't
fail (it's TEST mode!).