Got rid of explicit file existence check
This was causing problems. If the first git mv failed because the repo
was locked, the next iteration would fail because it looks for a file
that's not there and the 'git reset --hard' hasn't executed yet to reset
the local repo. If the file isn't where it should be let mv/git mv
report the error.