Make one variantbam job sleep a little before running
Make the variantbam job for NORMAL bam sleep for 10 seconds before
running. Sometimes (not often but often enough to be mildly annoying) it
seems that the two variantbam jobs try to create the same directory at
the same time, one will fail and retry, but the failure will move the
job file to failed-jobs and the problem is: currently there is no way to
move the job file OUT of failed-jobs is the workflow step retries
successfully.